requirements.txt
python==3.8.10
pytorch==1.11.0
numpy==1.22.3
scikit-learn==1.0.2
scipy==1.8.0
faiss-gpu==1.7.2